Modified client file structure. Added general components. Added a drawer to acces user's information
