--- client/src/MyBooksPage.css
@@ -1,185 +1,0 @@
-.my-books-page {
-    background-color: lightblue;
-    position: absolute;
-    overflow: hidden;
-
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-
-    width: 100vw;
-    min-height: 100vh;
-}
-
-.my-books-page header {
-    background-color: black;
-    height: 70px;
-    width: 100vw;
-    padding-left: 3%;
-    padding-right: 3%;
-
-    position: fixed;
-    top: 0;
-    z-index: 1;
-
-    display: flex;
-    flex-direction: row;
-    justify-content: space-evenly;
-    align-items: center;
-}
-
-.my-books-page img {
-    border-radius: 100px;
-    height: 55px;
-    width: 55px;
-    cursor: pointer;
-}
-
-.my-books-page .my-books-header img:hover {
-    box-shadow: 1px 1px 3px white;
-}
-
-.my-books-page .my-books-header button {
-    background-color: white;
-    color: black;
-
-    border: none;
-    border-radius: 100px;
-    padding: 10px 10px;
-    font-size: 20px;
-
-    cursor: pointer;
-}
-
-.my-books-page .my-books-header button:hover {
-    opacity: 0.8;
-}
-
-.my-books-page .my-books-header button:active {
-    opacity: 0.6;
-}
-
-.my-books-page .books-container {
-    margin-top: 71px;
-    display: grid;
-    grid-template-columns: 1fr;
-    column-gap: 20px;
-    row-gap: 20px;
-}
-
-.my-books-page .book-box {
-    background-color: white;
-    padding-bottom: 150px;
-    height: 600px;
-    width: 320px;
-
-    box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2),
-        0 6px 20px 0 rgba(0, 0, 0, 0.19);
-}
-
-.my-books-page .book-thumbnail {
-    display: flex;
-    justify-content: center;
-    height: 100%;
-}
-
-.my-books-page .book-infos {
-    display: flex;
-    flex-direction: column;
-    margin-left: 20px;
-    margin-right: 20px;
-}
-
-.my-books-page .book-title {
-    width: 100%;
-    overflow: hidden;
-    white-space: nowrap;
-    text-overflow: ellipsis;
-}
-
-.my-books-page .book-author {
-    width: 100%;
-    overflow: hidden;
-    white-space: nowrap;
-    text-overflow: ellipsis;
-}
-
-.my-books-page .book-genre {
-    width: 100%;
-    overflow: hidden;
-    white-space: nowrap;
-    text-overflow: ellipsis;
-}
-
-.my-books-page .remove-button-container {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    margin-top: 10px;
-}
-
-.my-books-page .remove-button {
-    background-color: red;
-    color: black;
-
-    font-size: 18px;
-    width: 85%;
-
-    border: none;
-    border-radius: 20px;
-    padding: 5px 30px;
-
-    cursor: pointer;
-}
-
-.my-books-page .book-box button:hover {
-    opacity: 0.8;
-}
-
-.my-books-page .book-box button:active {
-    opacity: 0.6;
-}
-
-.my-books-page .pagination {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    justify-content: space-around;
-    position: relative;
-    top: 4%;
-}
-
-.my-books-page .pagination button {
-    background-color: white;
-    color: black;
-
-    margin-top: 7px;
-    padding: 5px 20px;
-
-    border: none;
-    border-radius: 20px;
-    font-size: 25px;
-
-    cursor: pointer;
-}
-
-.my-books-page .pagination button:hover {
-    opacity: 0.8;
-}
-
-.my-books-page .pagination button:active {
-    opacity: 0.6;
-}
-
-@media only screen and (min-width: 768px) {
-    .my-books-page .books-container {
-        grid-template-columns: 1fr 1fr;
-    }
-}
-
-@media only screen and (min-width: 1500px) {
-    .my-books-page .books-container {
-        grid-template-columns: 1fr 1fr 1fr 1fr;
-    }
-}
