+++ client/src/components/Header.css
@@ -1,0 +1,47 @@
+header {
+    background-color: black;
+    height: 70px;
+    width: 100vw;
+    padding-left: 3%;
+    padding-right: 3%;
+
+    position: fixed;
+    top: 0;
+    z-index: 1;
+
+    display: flex;
+    flex-direction: row;
+    justify-content: space-evenly;
+    align-items: center;
+}
+
+img {
+    border-radius: 100px;
+    height: 55px;
+    width: 55px;
+    cursor: pointer;
+}
+
+.library-header img:hover {
+    box-shadow: 1px 1px 3px white;
+}
+
+.library-header button {
+    background-color: white;
+    color: black;
+
+    border: none;
+    border-radius: 100px;
+    padding: 10px 10px;
+    font-size: 20px;
+
+    cursor: pointer;
+}
+
+.library-header button:hover {
+    opacity: 0.8;
+}
+
+.library-header button:active {
+    opacity: 0.6;
+}
