+++ client/src/img/book-cover-placeholder.png
