+++ client/src/main.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App.jsx';
-import './Index.css';
+import './styles/Index.css';
 
 import { disableReactDevTools } from '@fvilers/disable-react-devtools';
