+++ client/src/pages/ForgotPasswordPage.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 
-import './ForgotPasswordPage.css';
+import '../styles/ForgotPasswordPage.css';
 
 const ForgotPasswordPage = () => {
     let URL;
