+++ client/src/pages/LibraryPage.jsx
@@ -1,13 +1,38 @@
 import React, { useState, useEffect } from 'react';
+import { useNavigate } from 'react-router-dom';
 
-import './LibraryPage.css';
-import bookPlaceholder from '../img/book-cover-placeholder.png';
-import profilePicture from '../img/profile_pic.jpg';
+import '../styles/LibraryPage.css';
+
+import Drawer from '../components/Drawer';
+import Header from '../components/Header';
+import BookContainer from '../components/BookContainer';
 
 function LibraryPage() {
+    const navigate = useNavigate();
+
     const [totalCount, setTotalCount] = useState(0);
     const [currentPage, setCurrentPage] = useState(1);
     const [itemsPerPage, setItemsPerPage] = useState(12);
+
+    const [isDrawerOpen, setIsDrawerOpen] = useState(false);
+    const handleDrawer = () => {
+        setIsDrawerOpen(!isDrawerOpen);
+    };
+    useEffect(() => {
+        document.body.style.overflow = isDrawerOpen
+            ? 'hidden'
+            : 'unset';
+
+        const booksContainer = document.querySelector(
+            '.books-container'
+        );
+        booksContainer.style.filter = isDrawerOpen
+            ? 'brightness(80%)'
+            : 'brightness(100%)';
+        booksContainer.style.pointerEvents = isDrawerOpen
+            ? 'none'
+            : '';
+    }, [isDrawerOpen]);
 
     let URL;
     if (`${import.meta.env.VITE_NODE_ENV}` === 'development') {
@@ -100,52 +125,20 @@ function LibraryPage() {
 
     return (
         <div className='library-page'>
-            <header className='library-header'>
-                <img
-                    className='library-logo'
-                    src={bookPlaceholder}
-                    onClick={() => window.open('/library', '_self')}
-                />
-                <button onClick={handlePreviousPage}>Previous</button>
-                <button onClick={handleNextPage}>Next</button>
-                <img
-                    className='profile-picture'
-                    src={profilePicture}
-                    onClick={() =>
-                        window.open('/user/my-books', '_self')
-                    }
-                />
-            </header>
+            <Drawer isDrawerOpen={isDrawerOpen} />
+            <Header
+                handlePreviousPage={handlePreviousPage}
+                handleNextPage={handleNextPage}
+                handleDrawer={handleDrawer}
+            />
             <div className='books-container'>
                 {librarypage.map((book) => (
-                    <div className='book-box' key={book._id}>
-                        <div className='book-thumbnail'>
-                            <img src={bookPlaceholder} />
-                        </div>
-                        <div className='book-infos'>
-                            <div className='book-title'>
-                                <strong>{book.name}</strong>
-                            </div>
-                            <div className='book-author'>
-                                Author: {book.author}
-                            </div>
-                            <div className='book-genre'>
-                                Genre: {book.genre}
-                            </div>
-                            <div>Year: {book.year}</div>
-                            <div>Pages: {book.pages}</div>
-                        </div>
-                        <div className='add-button-container'>
-                            <button
-                                className='add-button'
-                                onClick={() =>
-                                    handleAddToLibrary(book._id)
-                                }
-                            >
-                                Add to my library
-                            </button>
-                        </div>
-                    </div>
+                    <BookContainer
+                        book={book}
+                        onAction={handleAddToLibrary}
+                        actionText={'Add to my library'}
+                        key={book._id}
+                    />
                 ))}
             </div>
         </div>
