+++ client/src/pages/LogoutPage.jsx
@@ -1,0 +1,3 @@
+const LogoutPage = () => {};
+
+export default LogoutPage;
