+++ client/src/styles/ForgotPasswordPage.css
