+++ client/src/styles/MyBooksPage.css
@@ -1,0 +1,33 @@
+.my-books-page {
+    background-color: lightblue;
+    position: absolute;
+    overflow: hidden;
+
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+
+    width: 100vw;
+    min-height: 100vh;
+}
+
+.my-books-page .books-container {
+    margin-top: 71px;
+    display: grid;
+    grid-template-columns: 1fr;
+    column-gap: 20px;
+    row-gap: 20px;
+}
+
+@media only screen and (min-width: 768px) {
+    .my-books-page .books-container {
+        grid-template-columns: 1fr 1fr;
+    }
+}
+
+@media only screen and (min-width: 1500px) {
+    .my-books-page .books-container {
+        grid-template-columns: 1fr 1fr 1fr 1fr;
+    }
+}
